Break tests out into separate files, to encourage more files
